+++ typesense_codegen/Cargo.toml
@@ -11,7 +11,7 @@ serde = "^1.0"
 serde_derive = "^1.0"
 serde_json = "^1.0"
 url = "^2.2"
-uuid = { version = "^1.0", features = ["serde", "v4"] }
+uuid = { version = "^1.0", features = ["serde", "v4", "js"] }
 [dependencies.reqwest]
 version = "^0.11"
 features = ["json", "multipart"]
