+++ typesense_codegen/src/lib.rs
@@ -1,10 +1,5 @@
 #[macro_use]
 extern crate serde_derive;
 
-extern crate reqwest;
-extern crate serde;
-extern crate serde_json;
-extern crate url;
-
 pub mod apis;
 pub mod models;
